+++ GameEngine/GameEngine/Room.h
@@ -42,7 +42,7 @@ private:
 
 	void init();
 	List<std::string> parseValues(std::string line);
-	int parseInt(std::string line);
+	float parseFloat(std::string line);
 	bool collision(Object * obj1, Object * obj2);
 	void sortPlace(Object * obj, int index);
