+++ GameEngine/GameEngine/TerrainMap.cpp
@@ -19,14 +19,14 @@ void TerrainMap::draw(Object * player){
 	shader->bind();
 	Model* model = Res::get(ModelType::MODEL_SQUARE_CENTERED);
 	model->bind();
-	float xOffset = tiles.size() / 2;
-	float yOffset = tiles.get(0).size() / 2;
-	float xLoc = (player->pos[0] / 100) + (tiles.size()/2);
-	float yLoc = -(player->pos[1] / 100) + (tiles.get(0).size()/2);
-	int offset = 10;
+	float xOffset = tiles.size() / 2.0f;
+	float yOffset = tiles.get(0).size() / 2.0f;
+	float xLoc = (player->pos[0] / 100.0f) + (tiles.size()/2.0f);
+	float yLoc = -(player->pos[1] / 100.0f) + (tiles.get(0).size()/2.0f);
+	float offset = 10.0f;
 	
-	for (int i = yLoc - offset; (i < yLoc + offset) && (i < tiles.get(0).size()); i++) {
-		for (int k = xLoc - offset; (k < xLoc + offset) && (k < tiles.size()); k++) {
+	for (float i = yLoc - offset; (i < yLoc + offset) && (i < tiles.get(0).size()); i++) {
+		for (float k = xLoc - offset; (k < xLoc + offset) && (k < tiles.size()); k++) {
 			if (i < 0 || k < 0) {
 				continue;
 			}
@@ -87,7 +87,7 @@ List<int> TerrainMap::parseInts(std::string str)
 {
 	List<int> vals;
 	std::string current = "";
-	for (int i = 0; i < str.size(); i++) {
+	for (unsigned int i = 0; i < str.size(); i++) {
 		if (str[i] == ' ') {
 			std::stringstream stream(current);
 			int val;
